requirements.txt
streamlit
opencv-python
mediapipe
pillow
openai
